Drop unused / typo timestamp column
It was meant to be t.timestamps :null => false to magically
add created_at and updated_at columns but without the 's', it
expects a column name.

In rails 4.2, 5.0, 5.1, and 5.2, this line is ignored without a column name.
In rails 6.0, it raises: Missing column name(s) for timestamp.

It's been 5 years, we didn't need timestamps up until now, so we'll just drop it.